moving card component to a specific dir
